Medical Image Modalities
Medical imaging is the technique and process of creating visual representations of the interior of a body for clinical analysis and medical intervention, as well as visual representation of the function of some organs or tissues (physiology). Medical imaging seeks to reveal internal structures hidden by the skin and bones, as well as to diagnose and treat disease. Medical imaging also establishes a database of normal anatomy and physiology to make it possible to identify abnormalities. Although imaging of removed organs and tissues can be performed for medical reasons, such procedures are usually considered part of pathology instead of medical imaging.

May 15th, we will discuss and learn about diffrent medical image modalities and utilization. 
CT: Computed Tomography, also commonly referred to as a CAT scan, is a medical imaging method that combines multiple X-ray projections taken from different angles to produce detailed cross-sectional images of areas inside the body.  CT images allow doctors to get very precise, 3-D views of certain parts of the body, such as soft tissues, the pelvis, blood vessels, the lungs, the brain, the heart, abdomen and bones. CT is also often the preferred method of diagnosing many cancers, such as liver, lung and pancreatic cancers.
MRI: Magnetic Resonance Imaging  is a medical imaging technology that uses radio waves and a magnetic field to create detailed images of organs and tissues. MRI has proven to be highly effective in diagnosing a number of conditions by showing the difference between normal and diseased soft tissues of the body.
PET: Positron Emission Tomography is a nuclear imaging technique that provides physicians with information about how tissues and organs are functioning. PET, often used in combination with CT imaging, uses a scanner and a small amount of radiopharmaceuticals which is injected into a patient’s vein to assist in making detailed, computerized pictures of areas inside the body.
ultrasound: Diagnostic ultrasound, also known as medical sonography or ultrasonography, uses high frequency sound waves to create images of the inside of the body. The ultrasound machine sends sound waves into the body and is able to convert the returning sound echoes into a picture. Ultrasound technology can also produce audible sounds of blood flow, allowing medical professionals to use both sounds and visuals to assess a patient’s health.
X-ray: technology is the oldest and most commonly used form of medical imaging. X-rays use ionizing radiation to produce images of a person’s internal structure by sending X-ray beams through the body, which are absorbed in different amounts depending on the density of the material. In addition, included as “x-ray type” devices are also mammography,  interventional radiology, computed radiography, digital radiography and computed tomography (CT).  Radiation Therapy is a type of device which also utilizes either x-rays, gamma rays, electron beams or protons to treat cancer.